working on login bug
